LayoutTests/fast/regions/webkit-named-flow-event-remove-from-flow-expected.txt

Tests regionLayoutEvent

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

 
Flow content
PASS event.target.name is "article"
Removing region_2 from flow
PASS event.target.name is "article"
PASS successfullyParsed is true

TEST COMPLETE
